# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Configuration Management API",
    "version": "v1",
    "description": "Multi-tenant configuration management with versioning and access control"
  },
  "paths": {
...
    "/api/v1/ns/{namespace}/~navigate/": {
      "get": {
        "operationId": "navigate_root",
        "summary": "Navigate Root",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "recursive",
            "schema": {
              "default": false,
              "title": "Recursive",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigationSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Navigate tree from root.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~navigate/{path}": {
      "get": {
        "operationId": "navigate_path",
        "summary": "Navigate Path",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "recursive",
            "schema": {
              "default": false,
              "title": "Recursive",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NavigationSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Navigate tree at a specific path.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~search/": {
      "get": {
        "operationId": "search_root",
        "summary": "Search Root",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "types",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Types",
              "type": "array"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedTreeNavigationNodeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Search from namespace root.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~search/{path}": {
      "get": {
        "operationId": "search_path",
        "summary": "Search Path",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "types",
            "schema": {
              "items": {
                "type": "string"
              },
              "title": "Types",
              "type": "array"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedTreeNavigationNodeSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Search within a subtree.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~versions/{path}": {
      "get": {
        "operationId": "list_item_versions",
        "summary": "List Item Versions",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "default": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tagged_only",
            "schema": {
              "default": false,
              "title": "Tagged Only",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionHistoryResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "List all versions (metadata) for a config, template, or secret.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~get/{path}": {
      "get": {
        "operationId": "get_item",
        "summary": "Get Item",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "default": "latest",
              "title": "Version",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "reveal",
            "schema": {
              "default": false,
              "title": "Reveal",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "config": "#/components/schemas/ConfigSchemaExtended",
                      "folder": "#/components/schemas/FolderSchema",
                      "resolver": "#/components/schemas/ResolverSchema",
                      "secret": "#/components/schemas/SecretSchemaExtended",
                      "template": "#/components/schemas/TemplateSchemaExtended"
                    },
                    "propertyName": "node_type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConfigSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/TemplateSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/SecretSchemaExtended"
                    },
                    {
                      "$ref": "#/components/schemas/ResolverSchema"
                    },
                    {
                      "$ref": "#/components/schemas/FolderSchema"
                    }
                  ],
                  "title": "Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Get item details. For secrets, include ?reveal=true for decrypted content.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~diff/{path}": {
      "get": {
        "operationId": "diff_item",
        "summary": "Diff Item",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "from",
            "schema": {
              "default": "latest",
              "title": "From",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "to",
            "schema": {
              "default": "latest",
              "title": "To",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "to_path",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Path"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "reveal",
            "schema": {
              "default": false,
              "title": "Reveal",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponseSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Diff two versions of the same item (?from=, ?to=) or two items (?to_path=).\n\nReturns both sides for client-side diff rendering. For secrets, use\n?reveal=true to include decrypted content; otherwise decryption_required is set.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~delete/{path}": {
      "delete": {
        "operationId": "delete_item",
        "summary": "Delete Item",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "preview",
            "schema": {
              "default": true,
              "title": "Preview",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Soft-delete only this version (number or tag, e.g. 2 or stable)",
              "title": "Version"
            },
            "required": false,
            "description": "Soft-delete only this version (number or tag, e.g. 2 or stable)"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSchema"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Delete a tree item or a specific version. preview=true (default) is a dry run.",
        "tags": [
          "Tree"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~move/{path}": {
      "post": {
        "operationId": "move_item",
        "summary": "Move Item",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "config": "#/components/schemas/ConfigSchema",
                      "folder": "#/components/schemas/FolderSchema",
                      "resolver": "#/components/schemas/ResolverSchema",
                      "secret": "#/components/schemas/SecretSchema",
                      "template": "#/components/schemas/TemplateSchema"
                    },
                    "propertyName": "node_type"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ConfigSchema"
                    },
                    {
                      "$ref": "#/components/schemas/TemplateSchema"
                    },
                    {
                      "$ref": "#/components/schemas/SecretSchema"
                    },
                    {
                      "$ref": "#/components/schemas/ResolverSchema"
                    },
                    {
                      "$ref": "#/components/schemas/FolderSchema"
                    }
                  ],
                  "title": "Response"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSchema"
                }
              }
            }
          }
        },
        "description": "Move an item or folder subtree to a new path.",
        "tags": [
          "Tree"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationPayload"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorSchema": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {},
                "type": "array"
              }
            ],
            "title": "Error"
          },
          "audit_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Audit log event ID when the failure was recorded server-side",
            "title": "Audit Event Id"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorSchema",
        "type": "object"
      },
...
      "NavigationSchema": {
        "properties": {
          "item": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TreeNavigationNodeSchema"
              },
              {
                "type": "null"
              }
            ]
          },
          "children": {
            "items": {
              "$ref": "#/components/schemas/TreeNavigationNodeSchema"
            },
            "title": "Children",
            "type": "array"
          },
          "children_count": {
            "description": "Total number of child nodes (before pagination)",
            "title": "Children Count",
            "type": "integer"
          },
          "breadcrumbs": {
            "items": {
              "type": "string"
            },
            "title": "Breadcrumbs",
            "type": "array"
          },
          "is_leaf": {
            "title": "Is Leaf",
            "type": "boolean"
          }
        },
        "required": [
          "item",
          "children",
          "children_count",
          "breadcrumbs",
          "is_leaf"
        ],
        "title": "NavigationSchema",
        "type": "object"
      },
...
      "PagedTreeNavigationNodeSchema": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TreeNavigationNodeSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "count"
        ],
        "title": "PagedTreeNavigationNodeSchema",
        "type": "object"
      },
...
      "DeleteSchema": {
        "properties": {
          "delete": {
            "items": {
              "type": "string"
            },
            "title": "Delete",
            "type": "array"
          }
        },
        "required": [
          "delete"
        ],
        "title": "DeleteSchema",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ResolverAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "openid": "",
              "profile": "",
              "email": "",
              "groups": ""
            },
            "authorizationUrl": "http://localhost:8080/dex/auth",
            "tokenUrl": "http://localhost:8080/dex/token"
          }
        }
      }
    }
  },
  "servers": []
}